+++ api/src/main/java/org/allaymc/api/network/ProtocolInfo.java
@@ -4,6 +4,8 @@ import lombok.experimental.UtilityClass;
 import org.allaymc.api.utils.SemVersion;
 import org.allaymc.updater.block.BlockStateUpdater;
 import org.allaymc.updater.block.BlockStateUpdater_1_21_60;
+import org.allaymc.updater.item.ItemStateUpdater;
+import org.allaymc.updater.item.ItemStateUpdater_1_21_50;
 import org.cloudburstmc.protocol.bedrock.codec.BedrockCodec;
 import org.cloudburstmc.protocol.bedrock.codec.v776.Bedrock_v776;
 
@@ -37,6 +39,11 @@ public final class ProtocolInfo {
      * The currently used block state updater instance.
      */
     public static final BlockStateUpdater BLOCK_STATE_UPDATER = BlockStateUpdater_1_21_60.INSTANCE;
+
+    /**
+     * The currently used item state updater instance.
+     */
+    public static final ItemStateUpdater ITEM_STATE_UPDATER = ItemStateUpdater_1_21_50.INSTANCE;
 
     /**
      * The encoded version number of the block state version (without revision).